+++ R/aaa.R
@@ -1,3 +1,0 @@
-.onLoad <- function(libname, pkgname) {
-  #cpp_libproj_init_api()
-}